+++ db/db.go
@@ -407,6 +407,91 @@ func (db *DB) Backup(path string) error {
 	return dstDB.Close()
 }
 
+// Dump writes a consistent snapshot of the database in SQL text format.
+func (db *DB) Dump(w io.Writer) error {
+	if _, err := w.Write([]byte("PRAGMA foreign_keys=OFF;\nBEGIN TRANSACTION;\n")); err != nil {
+		return err
+	}
+
+	// Get a new connection, so the dump creation is isolated from other activity.
+	conn, err := open(fqdsn(db.path, db.dsn))
+	if err != nil {
+		return nil
+	}
+
+	// Get the schema.
+	query := `SELECT "name", "type", "sql" FROM "sqlite_master"
+              WHERE "sql" NOT NULL AND "type" == 'table' ORDER BY "name"`
+	rows, err := conn.Query([]string{query}, false, false)
+	if err != nil {
+		return err
+	}
+	row := rows[0]
+	for _, v := range row.Values {
+		table := v[0].(string)
+		var stmt string
+
+		if table == "sqlite_sequence" {
+			stmt = `DELETE FROM "sqlite_sequence";`
+		} else if table == "sqlite_stat1" {
+			stmt = `ANALYZE "sqlite_master";`
+		} else if strings.HasPrefix(table, "sqlite_") {
+			continue
+		} else {
+			stmt = v[2].(string)
+		}
+
+		if _, err := w.Write([]byte(fmt.Sprintf("%s;\n", stmt))); err != nil {
+			return err
+		}
+
+		tableIndent := strings.Replace(table, `"`, `""`, -1)
+		r, err := conn.Query([]string{fmt.Sprintf(`PRAGMA table_info("%s")`, tableIndent)}, false, false)
+		if err != nil {
+			return err
+		}
+		var columnNames []string
+		for _, w := range r[0].Values {
+			columnNames = append(columnNames, fmt.Sprintf(`'||quote("%s")||'`, w[1].(string)))
+		}
+
+		query = fmt.Sprintf(`SELECT 'INSERT INTO "%s" VALUES(%s)' FROM "%s";`,
+			tableIndent,
+			strings.Join(columnNames, ","),
+			tableIndent)
+		r, err = conn.Query([]string{query}, false, false)
+		if err != nil {
+			return err
+		}
+		for _, x := range r[0].Values {
+			y := fmt.Sprintf("%s;\n", x[0].(string))
+			if _, err := w.Write([]byte(y)); err != nil {
+				return err
+			}
+		}
+	}
+
+	// Do indexes, triggers, and views.
+	query = `SELECT "name", "type", "sql" FROM "sqlite_master"
+			  WHERE "sql" NOT NULL AND "type" IN ('index', 'trigger', 'view')`
+	rows, err = conn.Query([]string{query}, false, false)
+	if err != nil {
+		return err
+	}
+	row = rows[0]
+	for _, v := range row.Values {
+		if _, err := w.Write([]byte(fmt.Sprintf("%s;\n", v[2]))); err != nil {
+			return err
+		}
+	}
+
+	if _, err := w.Write([]byte("COMMIT;\n")); err != nil {
+		return err
+	}
+
+	return nil
+}
+
 // normalizeRowValues performs some normalization of values in the returned rows.
 // Text values come over (from sqlite-go) as []byte instead of strings
 // for some reason, so we have explicitly convert (but only when type