+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 4.6.0 (unreleased)
+- [PR #589](https://github.com/rqlite/rqlite/pull/589): Add restore command to CLI. Thanks @eariassoto.
 - [PR #588](https://github.com/rqlite/rqlite/pull/588): Abort transaction if load operation fails.
 - [PR #587](https://github.com/rqlite/rqlite/pull/587): Add expvar stats to store.
 - [PR #586](https://github.com/rqlite/rqlite/pull/586): rqlite CLI now supports command history. Thanks @rhnvrm.